+++ src/utils/snippetUtils.ts
@@ -1,6 +1,7 @@
 import * as vscode from "vscode";
 
 import { AssistantRecipe, Language } from "../graphql-api/types";
+import { filterImports } from "./dependencies/filter-dependencies";
 import { firstLineToImport, hasImport } from "./fileUtils";
 import {
   adaptIndentation,
@@ -23,9 +24,20 @@ export const insertSnippet = (
   );
   const decodedCode = decodeIndent(decodeFromBase64);
   const snippet = new vscode.SnippetString(decodedCode);
-  editor.insertSnippet(snippet, initialPosition);
 
-  for (const importStatement of recipe.imports) {
+  /**
+   * Filter imports that really need to be added.
+   */
+  const filteredImports = filterImports(
+    recipe.imports,
+    language,
+    editor.document
+  );
+
+  /**
+   * Insert the imports at the top of the file.
+   */
+  for (const importStatement of filteredImports) {
     if (!hasImport(editor.document, importStatement)) {
       const snippetString = new vscode.SnippetString(importStatement + "\n");
       const line = firstLineToImport(editor.document, language);
@@ -33,6 +45,17 @@ export const insertSnippet = (
       editor.insertSnippet(snippetString, position);
     }
   }
+
+  /**
+   * For each import we have insert, we also need to add another line
+   * and compute the final position. We just offset the lines.
+   */
+  const finalPosition = new vscode.Position(
+    initialPosition.line + filterImports.length,
+    initialPosition.character
+  );
+
+  editor.insertSnippet(snippet, finalPosition);
 };
 
 /**
@@ -59,7 +82,7 @@ export const deleteInsertedCode = async (
 
   await editor.edit((editBuilder) => {
     const previousDecodeFromBase64 = Buffer.from(
-      recipe.vscodeFormat || "",
+      recipe.presentableFormat || "",
       "base64"
     ).toString("utf8");
     const previousRecipeDecodedCode = adaptIndentation(
@@ -94,17 +117,17 @@ export const addRecipeToEditor = async (
   latestRecipeHolder: LatestRecipeHolder
 ): Promise<void> => {
   const currentIdentation = getCurrentIndentation(editor, initialPosition);
-
   if (currentIdentation === undefined) {
     return;
   }
 
-  const encodedCode = recipe.vscodeFormat;
+  const encodedCode = recipe.presentableFormat;
   const decodeFromBase64 = Buffer.from(encodedCode, "base64").toString("utf8");
   const decodedCode = adaptIndentation(
     decodeIndent(decodeFromBase64),
     currentIdentation
   );
+
   const latestRecipe = latestRecipeHolder.recipe;
 
   await editor.edit((editBuilder) => {
@@ -113,22 +136,24 @@ export const addRecipeToEditor = async (
      */
     if (latestRecipe) {
       const previousDecodeFromBase64 = Buffer.from(
-        latestRecipe?.vscodeFormat || "",
+        latestRecipe?.presentableFormat || "",
         "base64"
       ).toString("utf8");
       const previousRecipeDecodedCode = adaptIndentation(
         decodeIndent(previousDecodeFromBase64),
         currentIdentation
       );
+
       const previousCodeAddedLines = previousRecipeDecodedCode.split("\n");
       const lastLineAdded = previousCodeAddedLines.pop() || "";
-
       editBuilder.delete(
         new vscode.Range(
           initialPosition,
           new vscode.Position(
             initialPosition.line + previousCodeAddedLines.length,
-            lastLineAdded.length
+            previousCodeAddedLines.length > 1
+              ? lastLineAdded.length
+              : lastLineAdded.length + currentIdentation
           )
         )
       );